Move all source file to special folder.
